Make MeepleLayer event-aware
